requirements.txt
discord.py
python-dotenv
aiosqlite
pytz
aiofiles
uvicorn
matplotlib
fastapi